Backend/requirements.txt
MarkupSafe==3.0.2
PyJWT==2.10.1
PyYAML==6.0.1
StrEnum==0.4.15
aiofiles==23.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.95.0
feedparser==6.0.11
frozenlist==1.6.0
gitdb==4.0.12
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
jsonschema-specifications==2025.4.1
jsonschema==4.23.0
multidict==6.4.3
mutagen==1.47.0
narwhals==1.36.0
newsapi-python==0.2.6
numpy==2.2.5
openai==1.55.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pip==25.0.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pydantic-core==2.33.1
pydantic==1.10.22
pydeck==0.9.1
pydub==0.25.1
pytest-mock==3.14.0
pytest==8.3.5
python-dateutil==2.9.0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
realtime==2.4.3
referencing==0.36.2
requests==2.31.0
rpds-py==0.24.0
rsa==4.9.1
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.26.1
storage3==0.11.3
streamlit==1.44.1
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.21.1
websockets==14.2
yarl==1.20.0
